Cart

+++ AUSHOP/src/main/java/AUSHOP/Model/CartItem.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CartItem {
-	private Long productId;
+	private int productId;
 	private String name;
 	private int quantity;
 	private double price;
